player: Clean up and remove flipper logs.
