+++ VisualPinball.Unity/VisualPinball.Unity/Game/Player.cs
@@ -5,8 +5,6 @@ using Unity.Mathematics;
 using UnityEngine;
 using VisualPinball.Engine.Common;
 using VisualPinball.Engine.Game;
-using VisualPinball.Engine.Math;
-using VisualPinball.Engine.VPT.Ball;
 using VisualPinball.Engine.VPT.Flipper;
 using VisualPinball.Engine.VPT.Kicker;
 using VisualPinball.Engine.VPT.Plunger;
@@ -24,7 +22,6 @@ using VisualPinball.Unity.VPT.Plunger;
 using VisualPinball.Unity.VPT.Rubber;
 using VisualPinball.Unity.VPT.Surface;
 using VisualPinball.Unity.VPT.Table;
-using Random = UnityEngine.Random;
 
 namespace VisualPinball.Unity.Game
 {
@@ -35,10 +32,6 @@ namespace VisualPinball.Unity.Game
 		private readonly Dictionary<Entity, IApiHittable> _hittables = new Dictionary<Entity, IApiHittable>();
 		private readonly Dictionary<Entity, FlipperApi> _flippers = new Dictionary<Entity, FlipperApi>();
 		private FlipperApi Flipper(Entity entity) => _flippers.Values.FirstOrDefault(f => f.Entity == entity);
-
-		#if FLIPPER_LOG
-		public static StreamWriter DebugLog;
-		#endif
 
 		private Table _table;
 		private BallManager _ballManager;
@@ -111,10 +104,6 @@ namespace VisualPinball.Unity.Game
 			_player = gameObject.GetComponent<Player>();
 
 			World.DefaultGameObjectInjectionWorld.GetOrCreateSystem<StaticCollisionSystem>().Player = _player;
-
-			#if FLIPPER_LOG
-			DebugLog = File.CreateText("flipper.log");
-			#endif
 		}
 
 		private void Start()
@@ -171,66 +160,6 @@ namespace VisualPinball.Unity.Game
 				_tableApi.Plunger("Plunger001")?.Fire();
 				_tableApi.Plunger("Plunger002")?.Fire();
 			}
-		}
-
-		#if FLIPPER_LOG
-		private void OnDestroy()
-		{
-			DebugLog.Dispose();
-		}
-		#endif
-
-	}
-
-	public class DebugBallCreator : IBallCreationPosition
-	{
-		private float _x;
-		private float _y;
-
-		private readonly float _kickAngle;
-		private readonly float _kickForce;
-
-		public DebugBallCreator()
-		{
-			_x = -1;
-			_y = -1;
-		}
-
-		public DebugBallCreator(float x, float y)
-		{
-			_x = x;
-			_y = y;
-		}
-
-		public DebugBallCreator(float x, float y, float kickAngle, float kickForce)
-		{
-			_x = x;
-			_y = y;
-			_kickAngle = kickAngle;
-			_kickForce = kickForce;
-		}
-
-		public Vertex3D GetBallCreationPosition(Table table)
-		{
-			if (_x < 0 || _y < 0) {
-				_x = Random.Range(table.Width / 6f, table.Width / 6f * 5f);
-				_y = Random.Range(table.Height / 8f, table.Height / 2f);
-			}
-			return new Vertex3D(_x, _y, 0);
-		}
-
-		public Vertex3D GetBallCreationVelocity(Table table)
-		{
-			return new Vertex3D(
-				MathF.Sin(_kickAngle) * _kickForce,
-				-MathF.Cos(_kickAngle) * _kickForce,
-				0
-			);
-		}
-
-		public void OnBallCreated(PlayerPhysics physics, Ball ball)
-		{
-			// nothing to do
 		}
 	}
 }
